Update theme function to use cat instead of full file copies
Originally this function created a new alacritty.yml by copying a full
file with the dark theme configuration embedded in it. This meant tracking
three almost identical, large files. Now a new file is created by
concatenating the head and tail of the original file, with the theme
injected in the middle.