Update zshrc
- Remove bat as NULLCMD
- Set postgres environment variables dynamically

Homebrew installs each major version of postgres in a self contained
directory, along with its version number. This results in psql (and various
other things) that rely on a concrete binary path breaking each time a
new version is installed. To avoid this, these variables are now set
dynamically each time a shell is started